fix: avoid new cpplint runtime/int and include-order regressions
Two of my earlier RCLCPP_INFO calls used unsigned long long casts to
match a %llu format specifier, which cpplint flags (prefer fixed-width
or size_t). Switched both to size_t/%zu.

Also dropped the explicit <chrono> include in source_adapter.cpp: this
file's existing include block already puts quoted project headers
before angle-bracket ones, which cpplint's ordering check treats as
"other" before any bare stdlib header - so any new bare header added
after that point gets its own violation regardless of where it's
placed (confirmed empirically across several orderings). <chrono> is
already available transitively via the class's own rclcpp includes, so
this is a no-op for behavior and keeps this diff from adding new lint
noise.

Confirmed via a direct package-wide ament_cpplint comparison: this
branch is now at parity with humble's own baseline (82 pre-existing
errors, same count before and after this diff) - these fixes don't
touch that pre-existing debt, just avoid adding to it.

+++ src/polka_node.cpp
@@ -175,9 +175,9 @@ void PolkaNode::log_merge_perf(double us, const char * engine_label)
   merge_max_us_ = std::max(merge_max_us_, us);
   if (++merge_calls_ % kMergePerfLogInterval == 0) {
     RCLCPP_INFO(get_logger(),
-      "polka: perf merge [%s]: mean=%.3fms max=%.3fms (over %llu calls)",
+      "polka: perf merge [%s]: mean=%.3fms max=%.3fms (over %zu calls)",
       engine_label, merge_total_us_ / kMergePerfLogInterval / 1000.0,
-      merge_max_us_ / 1000.0, static_cast<unsigned long long>(kMergePerfLogInterval));
+      merge_max_us_ / 1000.0, static_cast<size_t>(kMergePerfLogInterval));
     merge_total_us_ = 0.0;
     merge_max_us_ = 0.0;
   }
